Supports multi push providers
Exposing handling of received messages without starting the service when
you have a custom firebase service that handles multiple push providers
